+++ src/components/VirtualInstallation.jsx
